api/requirements.txt
Flask==3.0.3
Flask-Limiter==3.7.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==0.7.1
psycopg2-binary==2.9.1
SQLAlchemy==1.4.22
marshmallow==3.12.2
jupiter-python-sdk==0.0.2.0
base58==2.1.0
telebot==0.0.5
python-dotenv==1.0.1
cryptography==42.0.8
pysolana==0.2.2
base58==2.2.1
ratelimit==2.2.1